Break up Sidebar code.
