+++ packages/apps/composer-app/src/components/Separator/Separator.tsx
@@ -1,0 +1,7 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import React from 'react';
+
+export const Separator = () => <div role='separator' className='bs-px mli-2.5 bg-neutral-500/20' />;
