+++ packages/apps/composer-app/src/components/Separator/index.ts
@@ -3,5 +3,3 @@
 //
 
 export * from './Separator';
-export * from './Sidebar';
-export * from './OctokitProvider';
