+++ packages/apps/composer-app/src/components/Sidebar/Sidebar.tsx
@@ -2,197 +2,33 @@
 // Copyright 2023 DXOS.org
 //
 
-import {
-  ArrowLineLeft,
-  Article,
-  ArticleMedium,
-  Circle,
-  DotsThreeVertical,
-  EyeSlash,
-  GearSix,
-  Intersect,
-  PaperPlaneTilt,
-  Planet,
-  Plus,
-  Sidebar
-} from '@phosphor-icons/react';
-import React, { useCallback, useContext, useEffect, useState } from 'react';
-import { useNavigate, useParams, Link } from 'react-router-dom';
+import { ArrowLineLeft, GearSix, Intersect, Planet, Sidebar } from '@phosphor-icons/react';
+import React, { useContext, useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import { Tooltip } from '@dxos/react-appkit';
-import { observer, ShellLayout, Space, useClient, useIdentity, useQuery, useSpaces } from '@dxos/react-client';
+import { observer, ShellLayout, useClient, useIdentity, useSpaces } from '@dxos/react-client';
 import {
   Avatar,
   Button,
-  buttonStyles,
   DensityProvider,
   Dialog,
-  DropdownMenu,
-  DropdownMenuItem,
   ElevationProvider,
   getSize,
   Input,
-  ListItemEndcap,
   mx,
   ThemeContext,
-  TooltipContent,
-  TooltipRoot,
-  TooltipTrigger,
-  TreeBranch,
-  TreeItem,
-  TreeItemBody,
-  TreeItemHeading,
   TreeRoot,
   useId,
   useTranslation
 } from '@dxos/react-components';
-import { TextKind } from '@dxos/react-composer';
 import { PanelSidebarContext, useShell } from '@dxos/react-ui';
 
 import { ComposerDocument } from '../../proto';
-import { abbreviateKey, getPath } from '../../router';
+import { getPath } from '../../router';
 import { useOctokitContext } from '../OctokitProvider';
-
-const DocumentTreeItem = observer(({ document, linkTo }: { document: ComposerDocument; linkTo: string }) => {
-  const { t } = useTranslation('composer');
-  const { docKey } = useParams();
-  const active = docKey === document.id;
-  const Icon = document.content.kind === TextKind.PLAIN ? ArticleMedium : Article;
-  return (
-    <TreeItem>
-      <TreeItemHeading asChild>
-        <Link
-          to={linkTo}
-          className={mx(buttonStyles({ variant: 'ghost' }), 'is-full text-base p-0 font-normal items-start gap-1')}
-          data-testid='composer.documentTreeItemHeading'
-        >
-          <Icon weight='regular' className={mx(getSize(4), 'shrink-0 mbs-2')} />
-          <p className='grow mbs-1'>{document.title || t('untitled document title')}</p>
-          <ListItemEndcap className='is-6 flex items-center'>
-            <Circle
-              weight='fill'
-              className={mx(getSize(3), 'text-primary-500 dark:text-primary-300', !active && 'invisible')}
-            />
-          </ListItemEndcap>
-        </Link>
-      </TreeItemHeading>
-    </TreeItem>
-  );
-});
-
-const SpaceTreeItem = observer(({ space }: { space: Space }) => {
-  const documents = useQuery(space, ComposerDocument.filter());
-  const { t } = useTranslation('composer');
-  const navigate = useNavigate();
-  const shell = useShell();
-  const { spaceKey, docKey } = useParams();
-  const identity = useIdentity();
-  const hasActiveDocument = !!(docKey && documents.map(({ id }) => id).indexOf(docKey) >= 0);
-
-  const handleCreate = useCallback(async () => {
-    const document = await space.db.add(new ComposerDocument());
-    return navigate(getPath(space.key, document.id));
-  }, [space, navigate]);
-
-  const handleViewInvitations = async () => shell.setLayout(ShellLayout.SPACE_INVITATIONS, { spaceKey: space.key });
-
-  const handleHideSpace = () => {
-    if (identity) {
-      const identityHex = identity.identityKey.toHex();
-      space.properties.members = {
-        ...space.properties.members,
-        [identityHex]: {
-          ...space.properties.members?.[identityHex],
-          hidden: true
-        }
-      };
-      if (spaceKey === abbreviateKey(space.key)) {
-        navigate('/');
-      }
-    }
-  };
-
-  const [open, setOpen] = useState(spaceKey === abbreviateKey(space.key));
-
-  useEffect(() => {
-    spaceKey === abbreviateKey(space.key) && setOpen(true);
-  }, [spaceKey]);
-
-  return (
-    <TreeItem
-      collapsible
-      open={open}
-      onOpenChange={setOpen}
-      slots={{
-        root: { className: 'mbe-2' },
-        ...(hasActiveDocument &&
-          !open && { openTriggerIcon: { weight: 'fill', className: 'text-primary-500 dark:text-primary-300' } })
-      }}
-    >
-      <div role='none' className='flex mis-1 items-start'>
-        <TreeItemHeading
-          className='grow break-words pbs-1.5 text-sm font-medium'
-          data-testid='composer.spaceTreeItemHeading'
-        >
-          {(space.properties.name?.length ?? 0) > 0 ? space.properties.name : space.key.truncate()}
-        </TreeItemHeading>
-        <TooltipRoot>
-          <TooltipContent className='z-[31]' side='bottom'>
-            {t('space options label')}
-          </TooltipContent>
-          <DropdownMenu
-            trigger={
-              <TooltipTrigger asChild>
-                <Button variant='ghost' data-testid='composer.openSpaceMenu' className='shrink-0 pli-1'>
-                  <DotsThreeVertical className={getSize(4)} />
-                </Button>
-              </TooltipTrigger>
-            }
-            slots={{ content: { className: 'z-[31]' } }}
-          >
-            <DropdownMenuItem asChild>
-              <Input
-                label={t('space name label')}
-                labelVisuallyHidden
-                value={space.properties.name ?? ''}
-                placeholder={space.key.truncate()}
-                onChange={({ target: { value } }) => (space.properties.name = value)}
-              />
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={handleViewInvitations} className='flex items-center gap-2'>
-              <PaperPlaneTilt className={getSize(4)} />
-              <span>{t('view space invitations label', { ns: 'os' })}</span>
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={handleHideSpace} className='flex items-center gap-2'>
-              <EyeSlash className={getSize(4)} />
-              <span>{t('hide space label')}</span>
-            </DropdownMenuItem>
-          </DropdownMenu>
-        </TooltipRoot>
-        <Tooltip content={t('create document label')} tooltipLabelsTrigger side='bottom' zIndex='z-[31]'>
-          <Button
-            variant='ghost'
-            data-testid='composer.createDocument'
-            className='shrink-0 pli-1'
-            onClick={handleCreate}
-          >
-            <span className='sr-only'>{t('create document label')}</span>
-            <Plus className={getSize(4)} />
-          </Button>
-        </Tooltip>
-      </div>
-      <TreeItemBody>
-        {documents.length > 0 && (
-          <TreeBranch collapsible={false}>
-            {documents.map((document) => (
-              <DocumentTreeItem key={document.id} document={document} linkTo={getPath(space.key, document.id)} />
-            ))}
-          </TreeBranch>
-        )}
-      </TreeItemBody>
-    </TreeItem>
-  );
-});
+import { Separator } from '../Separator';
+import { SpaceTreeItem } from './SpaceTreeItem';
 
 const DocumentTree = observer(() => {
   // TODO(wittjosiah): Fetch all spaces and render pending spaces differently.
@@ -316,9 +152,9 @@ const SidebarContent = () => {
                 </Button>
               </Tooltip>
             </div>
-            <div role='separator' className='bs-px mli-2.5 bg-neutral-500/20' />
+            <Separator />
             <DocumentTree />
-            <div role='separator' className='bs-px mli-2.5 bg-neutral-500/20' />
+            <Separator />
             {identity && (
               <div role='none' className='shrink-0 flex items-center gap-1 pli-3 plb-1.5'>
                 <Avatar
@@ -354,11 +190,7 @@ const SidebarToggle = () => {
   const { t } = useTranslation('os');
   const open = displayState === 'show';
   const button = (
-    <Button
-      data-testid='composer.toggleSidebar'
-      onClick={() => setDisplayState('show')}
-      className='p-0 is-[40px] shadow-md'
-    >
+    <Button data-testid='composer.toggleSidebar' onClick={() => setDisplayState('show')} className='p-0 is-[40px]'>
       <Sidebar className={getSize(6)} />
     </Button>
   );
